m42690/geometry: avoid aliasing when max bounding box is power of two
This commit fixes a bug where coding certain point clouds may
result in points being moved at the edge of the bounding box.
One such manifestation is the appearance of duplicate points in
a point cloud originally with unique point positions.

The maximum should be calculated as the lowest power of two minus
one that contains the bounding box.