geometry: replace bounding box with position and current tree depth
This optimisation reduces the size of the PCCOctree3Node structure
by replacing the per-node bounding box with the node's position (which
is the same as the bounding box minimums).  The contribution of the
bounding box maximums can be deduced from the initial bounding box
and the current tree depth.

Counters are used to track the transition from one depth to the next
during breadth-first traversal.