geometry: replace dynamic std::deque with fixed-size ring buffer
Use of an std::deque can involve a large number of malloc operations
as the internal sections of the queue are repeatedly allocated and
released during the geometry coding loop.

This commit provides a ring buffer implementation that uses the
worst-case octree node count (ie, ~number of input points) as
a fixed-size buffer.