geometry/encoder: use counting sort to reduce overhead
This commit eliminates the temporary fifo used in partitioning each node
into child nodes.  Rather than using independent sorting passes to apply
a binary partition to each of the x, y, and z axis, a counting sort
permits all eight output positions to be calculated with a single pass,
and a further pass to reorder the elements.