m42538/recolour: implement recolouring method of m42141 option 2
This commit eliminates the blending of point attribute values during
the attribute transfer process used in geometry quantisation by forcing
the blending weight to be 0|1 (m42141).

m42538 also recommends to disable the bounded search of the closest
attribute from the contributing set by setting searchRange=0.  This is
reflected by a change to the software defaults and configuration.