m42690/geometry: omit leaf node point count when mergeDuplicatePoints = 1
In cases where all point positions are unique (either due to input or
due to mergeDuplicatePoints = 1), there is no need to signal the number
of points in each leaf node (since it is always 1).

This commit adds a header flag to indicate this case, and modifies the
encoder and decoder to omit the point count signalling when
mergeDuplicatePoints = 1.