build: explicitly list tmc3 source files in CMakeLists.txt
In order to avoid obscure broken builds when switching branches
or otherwise adding/removing files, this commit removes the use
of wild cards for source files in the tmc3 directory.

To add files to or remove from the build, the tmc3/CMakeLists.txt
must be modified, thereby permitting the build system to detect
changes in the file list.