cfg: remove pre-generated configuration files
The provision of these configuration files causes confusion when they
are not kept in sync with the .yaml files -- either due to modification
of the .yaml files without regenerating the .cfg files, or due to
direct modification of the .cfg files themselves.

Please see cfg/README.md for further details on how to generate
configuration files for the encoder/decoder.