m43662/geometry: replace neighb=0 with sibling neighbour constraint
In the contextualisation of occupancy coding using neighbour patterns,
this commit replaces the case when neighbour contextualisation is
disabled with a constraint that restricts neighbour contextualisation
only examine the direct siblings of a node.

The restriction ensures that at most three neighbours may be present
for a node.

The configuration option "neighbourContextualisation=1" is replaced with
"neighbourContextRestriction=0".