m43953/hls: implement sps, gps, aps syntax structures and tlv fileformat
This commit rewrites the codec high-level syntax:

 - the bitstream is divided into "bricks" (akin to an AVC/HEVC
   slice/tile).

 - sequence, geometry and attribute parameter sets describe the coding
   parameters in use generally and for a specific brick.

 - marshalling the bitstream payloads to a file format is achieved using
   a type-length-value encoding scheme.

Additionally, the triSoup bitstream scale and translation values have
been unified with (replaced by) the octree counterparts.  For
compatibility, existing command line parameters continue to function
as before.

NB: this commit does not incorporate flexibility in the decoding order.
The decoder requires the bitstream to be presented in a fixed order.