m43953/hls: add more flexible decoder process
This commit allows individual payloads to be inserted into the decoder
rather than the decoder expecting to decode each payload in a fixed
order.

When an output pointcloud is ready, the decoder will call a user
supplied callback with the conforming output pointcloud.  The user
may choose to rescale this point cloud as necessary.