geom/m44750: occupancy context maps with on-the-fly update
Bitwise geometry occupancy coding uses a mapping table to select entropy
contexts.  This commit employs a dynamic mapping which is updated after
each coding operation, replacing the previous static mapping tables.

NB: the proposed version used a context with a halving period
(max_count) of 64 symbols.  However, this conflicts with another
adoption (512 symbols), and a wholesale replacement of the arithmetic
codec and context model.  To resolve the conflict, the existing halving
period (128) is used.