geom/m43600: intra geometry occupancy prediction
This commit provides a method to predict the child occupancy bits of a
node based on the node's 26 neighbours.  The prediction is used to
contextualise coding of each occupancy bit.

This tool requires the use of the occupancyAtlas for neighbour lookup.
NB: a restriction in the current implementation requires that the
atlas size is at most 8³.

intra_pred_max_node_size_log2: 6