octree/m43602: perform recolouring after geometry processing
This commit prepares for the possibility that geometry coding itself is
lossy.  Instead of performing recolouring just after input quantisation,
it is deferred until after geometry processing.

The recolouring functions are updated to handle differences in
pointcloud scales rather than requiring the target pointcloud to be
scaled identically to the source.