geom/m44752: false occupancy aware neighbour determination
The neighbour pattern for a node indicates occupied neighbouring nodes.

This commit updates the definition of occupancy for the (x-1), (y-1),
and (z-1) neighbours to be occupied if the neighbour's adjoining
children (ie sharing a face with the current node) are occupied.

In the 1-D figure below, at a depth d, the node 'X' has two occupied
neighbours 'R' and 'L'.  When determining the neighbour pattern of X,
the (x-1) state of the neighbour's adjoining child ('r') is used
instead.  For completeness, nodes marked with '?' are not causally
available.

  d+0:   R|X|L
  d+1:  Rr|??|??