cfg/m52525: set sequence bounding box for cat3 multi-frames sequences
When coding inter sequences with zero-motion prediction, the encoder's
default behaviour to compute a bounding box and local origin for each
frame causes the prediction to be misaligned.

This commit forces the bounding box and origin of all frames in the
sequence to be identical, thereby aligning the origins of each
frame and its reference.