attr/m47399: slice level signalling of QP deltas
This commit permits signalling a per-slice delta QP for luma and
chroma attribute components.  The delta is relative to the base
value derived from the APS.

NB: while syntax support and derivation processes are provided,
this commit does not contain any method to control the delta
QP value at the encoder.