geom/m52400: disable planar mode when quantisation can interfere
This addresses a conflict between the planar mode, which is determined
when a node is created as a child of its parent, and geometry
quantization, which is applied when a node's occupancy is computed.
This difference in processing may lead to a planar mode being used that
does not correspond to the quantized geometry.

This adoption disables the planar mode for octree nodes whose children
will perform quantization.