geom/m52343: update implicit qtbt rules for angular mode
For cat3-frame sequences, the implicit partitioning of a non-cubic
octree (qtbt) can be detrimental to the performance of the angular
contextualisation mode.  By signalling additional qtbt parameters,
it is possible to control the placement of narrower (ie, vertical)
nodes to occur deeper in the tree.  The additional parameters, are
determined by the encoder and signalled in the GPS.

NB: particular placements may cause the number of tree levels to
be greater than log2(maxNodeSize).