nanoflann: fix splitting axis determination for integer DistanceType
The computation of the largest dimension for splitting uses a greater-
than epsilon comparison assuming that a floating point DistanceType is
used.  The result is a greater-or-equal-than comparison.

However, with an integral DistanceType, epsilon is zero and a strict
greater-than comparison leads to an incorrect tree construction.

This commit changes the comparison to be always greater-or-equal-than.