attr/m52524: automatic calculation of LoD squared distance parameter
This is a non-normative method to determine a squared distance (dist2)
parameter for use in LoD attribute coding without any manual tuning.

An initial dist2 estimate is generated using an octree representation of
the point cloud to find the tree depth at which there number of internal
nodes is equal to a quarter of the input points.  A refinement process
performs a trial LoD construction to find a better dist2 estimate that
results in the desired ratio of points at the first level of detail.