slice/m52337: add uniform square partitioning method
This replaces the uniform geometry partitioning method:  First the input
point cloud is partitioned into uniformly square slices based on the
minimum axis.  Then slices are merged or split using 4n adjacency
information.  Merges are performed with the slice having the lowest
point count of the four neighbours.  Splits are processed according to
the direction of the next ordered slice.