predgeom/m53618: add angular predictive coding mode
This applies to point clouds acquired using a rotating scanner with one
or more capture elements scanning at a constant elevation. Such a model
covers, for instance, spinning LiDaR scanners.

This modifies the predictive geometry coder to code the location of a
point by first signalling a quantised spherical co-ordinate which is
predicted using the existing prediction methods.  The spherical
co-ordinate is converted to a Cartesian co-ordinate and a secondary
quantised residual applied to correct for any modelling error.

The encoder’s non-normative tree construction operates by assigning
input points in input order to a per-laser sub-tree.