enc: maintain input point order in quantizePositionsUniq
The original codec design did not care about the point order, and
therefore there was no harm in quantizePositionsUniq reordering points.
However, subsequent experiments on latency require maintaining the input
order.

This updates quantizePositionsUniq to maintain the original input order.
A side effect of this may be increased memory usage.