attr/m52331: partition scalable lod using distance to centroid
When generating LoDs with scalable coding, this partitions the current
level of detail according to the nearest point from the centroid of
occupied voxels in each 2x2x2 volume.  If there are two or more nodes
which are nearest to the centroid, the first or the last node in the
Morton order is chosen according to the detail level.