geom/m53390: explicitly signal qtbt partitioning
Rather than use an implicit process with various tweaks to tune the
process for various modes, this replaces the implicit process with
explicit signalling of the partitioning structure.  The encoder derives
the qtbt parameters according to the existing process and signals these
to the decoder.

By signalling the partitioning structure, it is no longer necessary to
explicitly signal the non-square root node size; this is instead derived
from the partitioning structure working up from the minimum node size.