geom/m53684: signal number of octree entropy streams
The original assumption that the octree node size is equivalent to the
number of tree levels remaining is no longer true.  This converts the
signalling of entropy streams to use the number of streams rather than
the node size at which they start.

A side-effect of this signalling is that the feature is now compatible
with trisoup.