hls/m53685: move geom_num_points_minus1 to slice footer
To aid encoders that must signal the exact number of points in the slice
which may only be known after encoding, the point count is moved to the
end of the data unit.

Decoders that wish to use the exact value can obtain this from the end
of the data unit, or may a weaker upper bound defined by the level limit
if waiting for the end of data unit is not an option.