hls/m53683: make tile inventory relative to sps origin
Tile inventory metadata should be relative to the sequence level origin.
This removes the ambiguity as to the frame of reference (global or
sequence) and reduces the overhead of signalling global tile positions.

In order to make the tile inventory stand alone, the sps origin
information is duplicated in the inventory.