hls/m53683: signal sps bounding box at conformance scale
Previously, it was required to comprehend the source scale factor in
order to apply the sequence bounding box to the output point cloud (or
conversely, to apply the output point cloud to the sequence bounding
box).

This change signals the sequence bounding box in the world-coordinate
system of the decoded output without applying any scaling.