scripts: use safe yaml loading in cfg-gen.pl
The YAML module interprets yaml tags by blessing the node with a
corresponding perl class.  For untrusted sources, this is unsafe and
new versions of the module disable this behaviour by default.

An explicit class has been implemented to handle the conditional yaml
tag without requiring the unsafe behaviour to be enabled.