attr/m54668: use look-up tables for lod generation
Proceeding in Morton code order of the input points at the current LoD
resolution, partitioning is performed on 2^7-cubed volumes. Each volume
is notionally divided into cubic regions of a power-of-two search
resolution.

For each point within the volume, a look-up table is used to test the
occupied state of the regions neighbouring the region containing the
current point.

Each occupied region is searched in turn for points previously promoted
to the next level of detail.

The current point is assigned to the current level of detail if a
neighbouring point is found that is closer than the level-of-detail
squared distance parameter. Otherwise, the current point is promoted to
the next level of detail.