hls/m54705: code potentially large ue(v) values using u(ue(v))
The following values have a potential to be inefficiently coded
as ue(v):
 - sps_bounding_box_offset, sps_bounding_box_size
 - geom_box_origin
 - attr_region_origin, attr_region_size_minus1
 - tile_origin, tile_size_minus1, ti_origin_xyz
 - num_unique_segments_minus1

To reduce the size of headers, these values are coded as u(n), where
the number of bits, n, is coded using ue(v) and shared between various
syntax elements.  In all cases, the value of n is not permitted to be
zero, so a _minus1 form of signalling is used.