attr/m53672: predict last coefficient component in lifting transform
This performs coefficient domain prediction of the last coefficient
component in three component attribute coding using the lifting
transform.  The last component is predicted according to a model of the
second coefficient component coeff[3] += coeff[2] * a.

The prediction coefficient model, a, is determined once per level of
detail and restricted to {-1,0,+1}.  A non-normative encoder process
determines values of a by correlating the sign of the second and third
coefficients.