trisoup: fix decoder where octree may have more nodes than points
It is possible that trisoup code fewer points than the trisoup nodes
used generate them.  If the octree the fifo buffer is sized according to
the guarantee provided by geom_num_points_minus1 (the number of decoded
points), an overflow may occur.

This works around the issue by allocating the fifo buffer based upon the
level limit when trisoup is enabled.  Since the number of points must be
lower than the level limit, and if a node must generate at least one
point, this should be a sufficient condition.