trisoup: fix encoder minimum octree node size for small trees
It is possible that the geometry being encoded by trisoup has a
dimension that is smaller than the trisoup node size.  While this is not
a problem for trisoup, it can cause the encoder to mistrim the list of
per-level node sizes generated by qtbt.

To fix this and to ensure logical consistency with the decoder node
sizes, it is necessary that the root node size at least the size of the
trisoup node size.