enc: use indices rather than positions to extract cloud partitions
Point cloud quantisation generates a map of quantised position to source
index.  This map is used to select the input points that correspond to
a single slice.  However, in certain cases, this map is not usable for
this purpose.

To resolve this, the position map is replaced by an index array map.