geom: remove incorrect assertion that occupancySkip != 7
The encoder and decoder both contained an assertion that occupancySkip is
not equal to 7, indicating that no axes are coded for the current node.
The intent of this assertion is to ensure that qtbt splitting is correct.

However, the assertion is misplaced, since, in the presence of in-tree
geometry quantisation, it is possible that an axis is not coded that
would have been otherwise.  If the resulting node is not a leaf node,
it is possible that no axes are coded, triggering the assertion.

This commit moves the assertion to the generation of the axis splitting
list instead.

For example, if the unquantised tree is quantised by qp=8 to remove
one bit from each position axis, the node at lvl 2 has no occupancy
coded:

     unquantised         | quantised (qp=8)
     NS   codedAxes(Lvl) | ENS  codedAxes(Node)
 lvl xyz  xyz            | xyz  xyz
 ... ...  ...            | ...  ...
  1  223  111            | 112  111
  2  112  110            | 001  XX0 => 000, no axes coded
  3  002  001            | 001  001
  4  001  001            | 000  00X <- NB: terminated at lvl3

Where:
 - lvl: the tree level,
 - NS:  the nodeSizeLog2,
 - ENS: the effectiveNodeSizeLog2 after quantisation
 - codedAxes: bit mask indicating which axes are coded

NB: codedAxes is the ones-complement of occupancySkip.