geom: ensure consistency of child occupancy atlas after idcm
Accesses to the child occupancy atlas data are guarded by checking that
the parent node is occupied.  This avoids the need to reset the child
occupancy data when the atlas resets.

However, idcm nodes by virtue of the coding order are indicated to be
occupied by the parent node.  In order to comply with the assumption
(that occupied parent nodes have valid child occupancy data) it is
necessary to clear the child occupancy associated with the idcm node.