geom: use coded axis list to derive atlasShift|occupancySkip (enc)
This is a non-normative simplification.  Rather than derive atlasShift
and occupancySkip from differences in nodeSizeLog2, which is itself
derived from tree_lvl_coded_axis_list, use the list directly.  This
additionally avoids the need to track the parent node size to perform
the derivation.

This commit modifies the encoder only.