geom: remove pointless optimisation of atlasShift = 0
atlasShift = 0 is impossible except for the root node (it indicates that
no bits of occupancy are coded in the preceding tree depth).  It is
therefore pointless to optimise for that case during the derivation of
the node neighbour pattern.

This commit changes removes the special case and restores the original
behaviour of the three parameter get()/getWithCheck() methods.