geom: fix incorrect determination of child neighbours under qtbt
The occupancy atlas contains occupancy information for two different
levels of the octree:

 - The main atlas is used to identify which nodes in the current level
   are present (occupied).  This is stored in a compact form, with one
   byte representing the occupancy of eight nodes.  When expressed in
   Morton order, this is identical to the eight occupancy bits of each
   node in the previous level.

 - A second atlas is used to identify the eight occupancy bits of each
   node in the current level.

Qtbt permits in any given level of the tree, up to two of the axes to
not be coded.  In order to query adjacent nodes in the atlas it is
necessary to take into account what axes are coded.

This commit fixes an issue where the second atlas is queried using
the coded axis information about the previous level, rather than
the information pertaining to the current level.