geom: determine node's planar nature from occupancy
When the planar nature of a node was signalled for each occupied child
of a node, it was necessary for an encoder to detect any planes present.

This commit removes the redundant plane detector (which was, in effect,
determining occupancy twice) now that the planar mode is signalled at
the start of each node, and instead determines the planes from the
node's occupancy byte.

This also has the effect of fixing an issue in the previous plane
detector when qtbt and in-tree geometry quantisation cause an axis to
have a negative effective node size.