geom: fix two-point ordered idcm in presence of quantisation
In the encoder, when points are quantised, the following representation
is used for positions, illustrated for an instance where a node is being
idcm coded after quantisation at a previous tree level:

   |-----|      = previously coded bits
          |---| = nodeSizeLog2 (to be coded)
   ppppppqqqq00 = quantised point bits
          |-|   = effectiveNodeSizeLog2

where:
 - p represents unquantised bits,
 - q represents quantised bits,
 - 0 are zero bits that are not coded due to division by a power-of-two,
 - nodeSizeLog2 is the size of the current node prior to quantisation,
 - effectiveNodeSizeLog2 is the size of the node after quantisation.

This commit fixes the incorrect use of effectiveNodeSizeLog2 to access
position bits without taking into account the zero padding by
introducing a local representation that omits the zero padding.

By way of convenience, all idcm coding modes are updated to operate in
the same manner.  However, this does not extend to fixing incompatibilities
between angular idcm and quantisation.