geom: don't zero the child occupancy atlas
Access to the child occupancy map is guarded by checking the parent
occupancy map first.  It is therefore not necessary to default
initialise the array or clear it between updates.

However, the relaxation of the idcm eligibility constraint introduced an
explicit initialisation of the array to avoid encoder/decoder
mismatches.

Since the idcm changes do not affect the precondition for accessing the
child occupancy atlas, this indicated that another aspect is causing it
to be violated.

Now that these issues have been fixed, this commit removed the explicit
initialisation that just serves to mask logic errors.

As an added protection, if built with -DWITH_MEMCHECK=1 and run under
valgrind, instrumentation is added that further avoids false negatives.