geom: don't calculate sibling neighbour pattern at parent level
The sibling neighbour pattern for a node may either be calculated at the
parent tree level and stored in the octree node, or computed for each
node at the time of use based upon sibling (parent) occupancy stored in
the node structure.

Since the sibling occupancy is required to be stored, and since the full
six-neighbour occupancy is calculated at the time of use from the
occupancy map, removing the neighbour pattern from the node structure
will save a byte (possibly four, depending upon alignment requirements)
per node in the fifo.

This commit implements the change to calculate the neighbour pattern.
A subsequent commit will remove it from the octree node structure.