geom/m55491: modulate idcm eligibility using rotating pattern
Idcm eligibility may be limited to trade-off the compression gain from
the planar mode with the run-time efficiency of early termination.

Rather than using the probability of the planar mode to perform this
rate limiting, an alternative process is used that does not require
reading back the value of context variables.

At the start of decoding, a 32bit pattern is generated with each bit
representing 1/32% of nodes being eligible.  Each idcm eligible node
is confirmed against the pattern, which is then rotated.