attr/m55364: direct predictor parsing dependency removal
Explicitly signalling a direct predictor for points which are implicitly
determined to have a weak prediction introduces a parsing dependency
upon reconstructed attribute values.

To remove the dependency, the conditional syntax elements are removed
and instead, the direct prediction data is encoded in the lsbs of the
quantised coefficient value.  The implicit process is used during
attribute reconstruction, independently of parsing, to decode the
direct predictor information from the quantised coefficient value.

An aps flag, direct_avg_predictor_disabled_flag, controls the ability
for the average predictor to participate in the direct predictor
selection.