geom/m55387: compensate for in-tree quantisation in angular coding
The node position is used in angular contextualisation of direct
positions or the planar plane position.  However, when in-tree
quantisation is enabled, the node position is in a different (quantised)
domain compared to the lidar origin, resulting in failures to determine
appropriate contexts and an associated loss in coding performance
compared to disabling quantisation.

These failures are remedied by scaling the node position, node size and
child (mid node) size for use by the contextualisation processes.