geom/m55379: store maximum node position coordinate in planar buffer
The planar buffer stores the position of the previous node within a
plane.  Rather than storing a two dimensional value, the maximum of the
two position components is stored instead.

When performing contextualisation, instead of using the Manhattan
distance between the current position and the previous position, the
absolute difference of the maximum components is used instead.

This results in an almost halving of the planar buffer size.