attr: num_detail_levels is actually num_detail_levels_minus1
This updates the name of the syntax variable to match the spec.

A point cloud naturally has one level of detail (its native resolution).
Subsampling will generate additional levels of detail.  The spec and
software use this convention, however, the variable name used in the
software is misleading and actually corresponded to the number of
refinement layers.