attr/m55369: remove secondary bitdepth
The secondary bitdepth permits coding the first attribute component with
a different bitdepth to the other components.  This is used, for
instance, in YCgCo-R coding, where the chroma bitdepth is greater than
the luma bitdepth.

This adoption removes the secondary bitdepth.  In order to code YCgCo-R,
the bitdepth for the luma component is increased to match the chroma,
however the coded range is kept as is.