predgeom/m55376: signal minimum radius for angular prediction
The minimum radius is used in prediction mode 0 (zero prediction) when
angular coding is enabled.  The minimum value is coded in the slice
header.

NB: the encoder may be unable to determine the minimum value if the
multiple trees are used per slice due to a configured max points per
tree limit.  Zero is used in this cases.