hls/m55378: add per-slice signalling of angular origin
This adds support to signal the angular origin independently for each
slice.  When signalled at the per-slice, it permits to code lidar
acquired sequences, where the camera is in motion, in a coordinate
system other than the camera's.

While the encoder supports per-slice signalling, this is disabled in
the current implementation.