hls/m55576: replace gsh_tile_id with slice_tag
This generalises the interpretation of gsh_tile_id.  Since tiles are
purely metadata, marking a slice with a tileid is effectively tagging
the slice.  Since applications may have more specific requirements than
provided for in the tile inventory, the slice_tag provides a means to
mark slices when the tile inventory is not present.

The width of the slice_tag is signalled in the SPS.