attr: fix incorrect bounds in scalable neighbour search
The scalable neighbour search examines a range [p1,k1].  This search
is optimised by dividing the range into buckets and then testing the
bounding boxes of the corresponding buckets.

This commit fixes an issue where the last bucket in the range is
sometimes not examined, resulting in the range [p1,k1) being searched
instead.