attr/m56205: sort scalable predictors by L2 weight
This aligns the scalable nearest neighbour search with the non-scalable
version.  The neighbours must be re-sorted to ensure that the lowest
L2 weight is first in order to meet preconditions on subsequent weight
derivation.

The re-sorting must occur after the max_neigh_range pruning process,
since the pruning process discards predictors according to the pre-
sorted order.