attr: fix scalable extended inter neighbour search precondition
The extended inter lod neighbour search is only performed if fewer than
three neighbours were found in the initial search.  This commit fixes an
issue where if num_pred_nearest_neighbours_minus1 is less than 2, the
extended inter search is always performed irrespective of the number of
neighbours found in the initial search.