hls/m54699: derive FrameCtr from slice frame_ctr_lsb
This defines how the most significant bits of FrameCtr are derived from
the relative value of frame_ctr_lsb.

NB: a side-effect of these rules is that the first frame number may
be negative.

Additionally, the overloading of FrameCtr to indicate the suppression
of non-existent frames, is replaced by an explicit state variable.