predgeom/m55979: quantize azimuth residual according to radius
An adaptive quantization step size is used to quantize the angular
predictive geometry azimuth angle residuals.  The effect is to provide
a uniform quantization of circular arcs for any radius.

The use of an increased phi precision, requires that the scaling applied
by spherical coordinate conversion for attribute coding must be adjusted
to generate positions that meet the level constraint (21-bit).

The `attrSphericalMaxLog2` option permits the encoder to adjust the
derived scale factors assuming a particular lower maximum value (which
would ordinarily have been derived from the phi component).  The ctc
conditions are updated to use similar values to v12.