hls/m55576: remove octree entropy stream offsets
This removes the offsets present in the slice header used to indicate
the length of each entropy stream.

Entropy streams are concatenated, after flushing of the entropy codec at
the end of each stream.

This includes the adoption of m56205 that permits the decoder to
correctly locate the end of a chunk when chunked entropy streams are
enabled.

Two entropy streams (where the last chunk of A is terminated early):
     ... Stream A---->|<-Stream B ...
     ...........|yyyxx|bbbbbbbb......

are concatenated as:
     ........|yyybbbxx|bbbbb.........
                      ^ expected end of A