geom/m52521: signal quantization depth on a per-level basis
This allows an encoder to decide on the fly at what level to perform
geometry quantization.  Instead of signalling the depth at in the
slice header, a flag is signalled at each tree level indicating
whether node qp offsets are present in that level.  The flag is not
signalled in subsequent tree levels.