hls/m56810: use attribute scaling parameters for reflectance scaling
This replaces dual encoder and decoder option hack.reflectanceScale that
performed quantization (encoder) and scaling (decoder) of reflectance
data.  Instead the attribute parameter syntax is used to convey the
scaling information.

The new per-attribute options attrScale and attrOffset configure the
scaling.  NB: this is only supported for reflectance data currently.