cli: replace sequence resolution with unit lengths
The use of resolution (n points per metre) to describe input and output
point clouds is awkward.  For example, a point cloud might be described
as having a resolution of 1000 points per metre, but this dose not
describe whether the input unit is points or metres.

This commit changes the cli to require the unit length of the input
to be specified instead, ie 1 point = 1mm.  The unit is explicitly
configured too (metre vs dimension less).