enc: manipulate seq_geom_scale according to an external system
The encoded sequence scale is defined to be relative to an externally
defined coordinate system.

The encoder does not provide a means to directly control the value
of seq_geom_scale without affecting pre-scaling.

Instead, the externalScale parameter may be used to manipulate the
coded value of seq_geom_scale.  For example, externalScale=2
indicates that 1 sequence unit is equal to 2 external units,
causing seq_geom_scale to be halved.

This parameter has no effect if the source unit is metres.